% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_prediction.R
\name{predict_class}
\alias{predict_class}
\title{Predict FL B cell Group Assignments}
\usage{
predict_class(state_coef_df)
}
\arguments{
\item{state_coef_df}{A dataframe containing B cell state coefficients, predicted or otherwise generated, which are organized into samples
as rows and coefficients as columns. The columns must have the column names of c("INFM", "PDZ", "CMI"), in that order.}
}
\value{
A named list of class assignments.
}
\description{
Predicts the absolute class/group designation of samples based on their normalized state values.
}
\examples{
\donttest{
data("test_exp", envir = environment())
state_coef_df <- Coef_predictor(test_exp, nu = 0.1, batch_correct = FALSE)
predict_class(state_coef_df)
}
}
