% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBstate.R
\docType{data}
\name{signature_mat}
\alias{signature_mat}
\title{FL B cell Signature Matrix}
\format{
A data frame with 5 columns and 150 rows
}
\usage{
data(signature_mat)
}
\description{
A data frame representing linear defining gene signatures for each of the three B cell states
}
