% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBstate.R
\docType{data}
\name{test_exp}
\alias{test_exp}
\title{FL Bulk Expression Example Matrix}
\format{
A data frame with 91 columns and 150 rows
}
\usage{
data(test_exp)
}
\description{
A data frame representing gene expression of the 150 signature genes across 91 samples. The values are log2TPM+1 values.
}
